+++ plugins/fancyholograms/src/main/java/com/fancyinnovations/fancyholograms/main/FancyHologramsPlugin.java
@@ -16,6 +16,7 @@ import com.fancyinnovations.fancyholograms.commands.lampCommands.conditions.Holo
 import com.fancyinnovations.fancyholograms.commands.lampCommands.fancyholograms.BackupCMD;
 import com.fancyinnovations.fancyholograms.commands.lampCommands.fancyholograms.ConfigCMD;
 import com.fancyinnovations.fancyholograms.commands.lampCommands.hologram.*;
+import com.fancyinnovations.fancyholograms.commands.lampCommands.traits.InteractionTraitCMD;
 import com.fancyinnovations.fancyholograms.commands.lampCommands.traits.MultiplePagesTraitCMD;
 import com.fancyinnovations.fancyholograms.commands.lampCommands.types.*;
 import com.fancyinnovations.fancyholograms.config.FHConfiguration;
@@ -383,6 +384,7 @@ public class FancyHologramsPlugin extends JavaPlugin implements FancyHolograms {
 
         // hologram trait commands
         lamp.register(MultiplePagesTraitCMD.INSTANCE);
+        lamp.register(InteractionTraitCMD.INSTANCE);
     }
 
     private void registerListeners() {